Move the es fixture to conftest.py
